refactor: update types file names and such imports
