+++ packages/dapp/src/types/index.ts
@@ -1,6 +1,6 @@
-export * from '../const/linkMap';
-export * from './featureCardsRequest';
-export * from './gtag';
+export * from './featureCardsRequest.types';
+export * from './gtag.types';
 export * from './i18n.types';
 export * from './internal.types';
 export * from './userTracking.types';
+export * from './utm.types';
